+++ src/Mod/TechDraw/Gui/TechDrawHandler.h
@@ -22,9 +22,13 @@
 
 #pragma once
 
+#include <QKeyEvent>
+#include <QMouseEvent>
+
 #include <Gui/ToolHandler.h>
 
 #include <Mod/TechDraw/TechDrawGlobal.h>
+#include <Mod/TechDraw/App/DrawPage.h>
 
 namespace TechDrawGui
 {
@@ -36,6 +40,8 @@ public:
 
     TechDrawHandler();
     virtual ~TechDrawHandler();
+
+    std::list<Gui::InputHint> getToolHints() const override;
 
     void activate(QGVPage* vPage);
     void deactivate() override;
